+++ CHANGELOG.md
@@ -25,7 +25,8 @@ Changelog entries are grouped by type, with the following types:
 - Added embedding support to `OpenRouter` provider (delegates to OpenAI)
 - Added embedding support to `TogetherAI` provider (delegates to OpenAI)
 - Added embedding support to `Vercel` provider (delegates to OpenAI)
-- Added `openai_compatible` provider. (#76) by @Blankeos
+- Added `openai_compatible` provider.
+- Added 54 new OpenAI-compatible providers
 
 ### Changed